requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
sqlalchemy==2.0.23
qdrant-client>=1.12.0
openai>=1.12.0,<2.0.0
anthropic==0.42.0
libtmux==0.23.2
pydantic>=2.11.0
pydantic-settings>=2.7.0
anyio>=4.11.0
httpx>=0.27.0,<0.29.0
mcp>=1.18.0
aiofiles==23.2.1
python-dotenv==1.0.0
watchdog==3.0.0
alembic==1.13.0
websockets==13.0
tenacity==8.2.3
structlog==23.2.0
prometheus-client==0.19.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-groq>=0.0.1
requests==2.31.0
pyyaml==6.0.1
textual==0.47.1
rich==13.7.0